--- extension/styles.css
@@ -1,8 +1,0 @@
-.container {
-    margin: 1.5em;
-    text-align: center;
-}
-  
-.result {
-margin-top: 1em;
-}
